--- server/server.ts
@@ -1,37 +1,0 @@
-import express from 'express';
-import db from './config/connection.js';
-
-// Import the ApolloServer class
-import { ApolloServer } from '@apollo/server';
-import { expressMiddleware } from '@apollo/server/express4';
-
-// Import the two parts of a GraphQL schema
-import { typeDefs, resolvers } from './schemas/index.js';
-
-const server = new ApolloServer({
-  typeDefs,
-  resolvers,
-});
-
-// Create a new instance of an Apollo server with the GraphQL schema
-const startApolloServer = async () => {
-
-  await server.start();
-  await db();
-
-  const PORT = process.env.PORT || 3001;
-  const app = express();
-
-  app.use(express.urlencoded({ extended: false }));
-  app.use(express.json());
-
-  app.use('/graphql', expressMiddleware(server));
-
-  app.listen(PORT, () => {
-    console.log(`API server running on port ${PORT}!`);
-    console.log(`Use GraphQL at http://localhost:${PORT}/graphql`);
-  });
-};
-
-// Call the async function to start the server
-startApolloServer();
