--- server/src/routes/api/index.ts
@@ -1,1 +1,0 @@
-
