--- server/src/seeds/seed.ts
@@ -1,1 +1,0 @@
-
