+++ server/src/utils/auth.ts
